Push notifications
When the Arduino detects a manual lock or unlock (AKA a physical key
was used) it will send a message to the server via BT which will then
notify the PHP server to send a push notification to all registered
devices. When the knock sensor is completed knocks will be handled the
same way.